run tests, added input control

+++ dspace-api/src/main/java/org/dspace/handle/HandleClarinServiceImpl.java
@@ -10,8 +10,10 @@ package org.dspace.handle;
 import java.sql.SQLException;
 import java.util.List;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.Logger;
-import org.dspace.content.DSpaceObject;
+import org.dspace.authorize.AuthorizeException;
+import org.dspace.authorize.service.AuthorizeService;
 import org.dspace.content.MetadataFieldServiceImpl;
 import org.dspace.content.service.ItemService;
 import org.dspace.core.Context;
@@ -21,7 +23,7 @@ import org.dspace.handle.service.HandleClarinService;
 import org.dspace.handle.service.HandleService;
 import org.dspace.services.ConfigurationService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.util.ObjectUtils;
 
 /**
  * Additional service implementation for the Handle object in Clarin-DSpace.
@@ -47,6 +49,9 @@ public class HandleClarinServiceImpl implements HandleClarinService {
     @Autowired(required = true)
     protected ConfigurationService configurationService;
 
+    @Autowired(required = true)
+    protected AuthorizeService authorizeService;
+
     /**
      * Public Constructor
      */
@@ -64,12 +69,20 @@ public class HandleClarinServiceImpl implements HandleClarinService {
     }
 
     @Override
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public Handle createHandle(Context context, String handleStr, DSpaceObject dSpaceObject, String url)
-            throws SQLException {
-        String handleId = null;
+    public Handle findByHandle(Context context, String handle) throws SQLException {
+        return handleDAO.findByHandle(context, handle);
+    }
 
-        //Do we generate the new handleId generated or use entered handleStr by user?
+    @Override
+    public Handle createExternalHandle(Context context, String handleStr, String url)
+            throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
+        String handleId = null;
+        //Do we want to generate the new handleId or use entered handleStr by user?
         if (handleStr != null) {
             //we use handleStr entered by use
             handleId = handleStr;
@@ -80,28 +93,21 @@ public class HandleClarinServiceImpl implements HandleClarinService {
 
         Handle handle = handleDAO.create(context, new Handle());
 
-        //set handle depending on handleId
-        handle.setHandle(handleId);
-        //only if dspace object exists
-        if (dSpaceObject != null) {
-            handle.setDSpaceObject(dSpaceObject);
-            handle.setResourceTypeId(dSpaceObject.getType());
-        }
+        log.debug("Created new external Handle with handle " + handleId);
 
-        if (url != null) {
-            handle.setUrl(url);
-        }
-
-        handleDAO.save(context, handle);
-
-        log.debug("Created new Handle with handle " + handleId);
+        //set handle and url in created handle
+        setHandleAndUrlOfHandleObject(context, handle, handleStr, url);
 
         return handle;
     }
 
     @Override
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public void delete(Context context, Handle handle) throws SQLException {
+    public void delete(Context context, Handle handle) throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
         //delete handle
         handleDAO.delete(context, handle);
 
@@ -110,8 +116,12 @@ public class HandleClarinServiceImpl implements HandleClarinService {
     }
 
     @Override
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public void save(Context context, Handle handle) throws SQLException {
+    public void save(Context context, Handle handle) throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
         //save handle
         handleDAO.save(context, handle);
 
@@ -122,33 +132,35 @@ public class HandleClarinServiceImpl implements HandleClarinService {
     }
 
     @Override
-    @PreAuthorize("hasAuthority('ADMIN')")
     public void update(Context context, Handle handleObject, String newHandle,
-                       DSpaceObject dso, String newUrl)
-            throws SQLException {
-        //set all handle attributes
-        handleObject.setHandle(newHandle);
-        handleObject.setDSpaceObject(dso);
-        //if dspace object is null, set resource type id to null
-        if (dso != null) {
-            //resource type id is type od dspace object
-            handleObject.setResourceTypeId(dso.getType());
-        } else {
-            handleObject.setResourceTypeId(null);
+                       String newUrl)
+            throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
         }
-        if (newUrl != null) {
-            handleObject.setUrl(newUrl);
-        }
-
-        this.save(context, handleObject);
+        //set handle and url in handle
+        setHandleAndUrlOfHandleObject(context, handleObject, newHandle, newUrl);
 
         log.info(LogHelper.getHeader(context, "update_handle",
                 "handle_id=" + handleObject.getID()));
     }
 
     @Override
-    @PreAuthorize("hasAuthority('ADMIN')")
-    public void setPrefix(Context context, String newPrefix, String oldPrefix) throws SQLException {
+  //  @PreAuthorize("hasAuthority('ADMIN')")
+    public void setPrefix(Context context, String newPrefix, String oldPrefix) throws SQLException,
+            AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
+        //control, if are new and old prefix entered
+        if (ObjectUtils.isEmpty(newPrefix) && StringUtils.isBlank(newPrefix) &&
+                ObjectUtils.isEmpty(oldPrefix) && StringUtils.isBlank(oldPrefix)) {
+            throw new NullPointerException("Cannot set prefix. Required fields are empty.");
+        }
         //get handle prefix
         String prefix = handleService.getPrefix();
         //set prefix only if not equal to old prefix
@@ -158,6 +170,8 @@ public class HandleClarinServiceImpl implements HandleClarinService {
                 //prefix has not changed
                 throw new RuntimeException("error while trying to set handle prefix");
             }
+        } else {
+            throw new RuntimeException("Cannot set prefix. Entered prefix does not match with ");
         }
 
         log.info(LogHelper.getHeader(context, "set_handle_prefix",
@@ -218,5 +232,37 @@ public class HandleClarinServiceImpl implements HandleClarinService {
         return handlePrefix + (handlePrefix.endsWith("/") ? "" : "/") + handleSuffix.toString();
     }
 
+    /**
+     * Set handle and url of handle object.
+     * It is not possible to change internal handle to external handle or
+     * external handle to internal handle.
+     *
+     * @param context       DSpace context object
+     * @param handleObject       handle object
+     * @param newHandle     new string handle
+     * @param newUrl     new url
+     * @throws SQLException if database error
+     * @throws AuthorizeException if authorization error
+     */
+    private void setHandleAndUrlOfHandleObject(Context context, Handle handleObject, String newHandle,
+                                           String newUrl) throws SQLException, AuthorizeException {
+        //set handle
+        handleObject.setHandle(newHandle);
+        //if it is internal handle, do nothing with url
+        if (newUrl != null) {
+            //set url only if is not empty
+            //when you add null to String, it converts null to "null"
+            if (!(ObjectUtils.isEmpty(newUrl)) && !(StringUtils.isBlank(newUrl)) &&
+                newUrl != "null") {
+                handleObject.setUrl(newUrl);
+            } else {
+                throw new RuntimeException("Cannot change handle and url of handle object.");
+            }
+        }
 
+        this.save(context, handleObject);
+
+        log.info(LogHelper.getHeader(context, "Set handle and url of handle object.",
+                "handle_id=" + handleObject.getID()));
+    }
 }
